backend/requirements.txt
# Core Framework
fastapi==0.118.0
uvicorn[standard]
pydantic==2.11.9
python-multipart

# Google AI
google-genai==1.38.0

# Database
sqlalchemy==2.0.36
psycopg2==2.9.10

# Session Storage
redis==5.2.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.1
markdown==3.9
requests>=2.32.3
beautifulsoup4==4.12.3
python-magic==0.4.27
